// src/utils/enums/colors.ts
export enum Colors {
  CONVO_BG = '#f4f6f8',
  BG_GREY = '#f4f6f8',
  WHITE = '#ffffff',
  BLACK = '#000000',
  DIVIDER_GREY = '#ebebed',
  BORDERS_GREY = '#e0e0e5',
  HOVER_GREY = '#f7f7f8',
  CTA_BG_GREY = '#efeff1',

  PRIMARY_1 = '#33334f',
  PRIMARY_2 = '#6d6da7',
  PRIMARY_3 = '#dee2e6',
  PRIMARY_4 = '#d9d8e8',

  BLACK_70 = '#707084',
  BLACK_50 = '#9999a7',
  BLACK_40 = '#adadb9',
  BLACK_20 = '#d6d6dc',

  EMP_BLUE_1 = '#3654ff',
  EMP_BLUE_2 = '#2d47de',
  EMP_BLUE_3 = '#273fca',

  SUCCESS_GREEN_1 = '#12c457',
  SUCCESS_GREEN_2 = '#15ad50',
  SUCCESS_GREEN_3 = '#149847',
  CONFIRMATION_GREEN = '#27ae60',
  REVIEWS_STARS_GREEN = '#1ba756',

  WARN_RED_1 = '#e63b3b',
  WARN_RED_2 = '#d62f2f',
  WARN_RED_3 = '#b92222',
  ERROR_RED = '#eb5757',

  RED_BG = '#fad5d5',

  ALERT_ORANGE_1 = '#ff9c28',
  ALERT_ORANGE_2 = '#ec9126',
  ALERT_ORANGE_3 = '#d27b15',
  ORANGE = '#f66933',

  DEST_YELLOW_1 = '#f8d50d',
  DEST_YELLOW_2 = '#eac90e',
  DEST_YELLOW_3 = '#d4b607',

  ERROR_TEXT = '#f0441e',

  BUTTON_TEXT_COLOR = '#dcdcf3',

  HEADER_BG = '#292940',

  ORANGE_DARK = '#ff8a1e',
}